+++ .pnp.js
@@ -3575,7 +3575,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/pnpify", "virtual:4618db608ddd5ff4b385ac4f8bba9302fe80be01f8be4f2c09326b6232f308ff5592ae5490aee3e37d21fb1703c6a7a0990d92b6e5f86a74e640f011f12da48b#workspace:packages/berry-pnpify"],
             ["brfs", "npm:2.0.1"],
             ["buffer-loader", "npm:0.1.0"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["joi", "npm:13.7.0"],
             ["pnp-webpack-plugin", "npm:1.4.3"],
             ["raw-loader", "npm:0.5.1"],
@@ -3616,7 +3616,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/pnp", "workspace:packages/berry-pnp"],
             ["@berry/shell", "workspace:packages/berry-shell"],
             ["chalk", "npm:2.4.2"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["semver", "npm:5.6.0"],
             ["tmp", "npm:0.0.33"],
             ["yup", "npm:0.27.0"]
@@ -3637,7 +3637,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["agentkeepalive", "npm:4.0.2"],
             ["camelcase", "npm:5.3.1"],
             ["chalk", "npm:2.4.2"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["cross-spawn", "npm:6.0.5"],
             ["globby", "npm:8.0.2"],
             ["got", "npm:9.6.0"],
@@ -3810,7 +3810,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/plugin-constraints", "workspace:packages/plugin-constraints"],
             ["@berry/core", "workspace:packages/berry-core"],
             ["@berry/fslib", "workspace:packages/berry-fslib"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["inquirer", "npm:6.2.1"],
             ["lodash.get", "npm:4.4.2"],
             ["node-emoji", "npm:1.8.1"],
@@ -3828,7 +3828,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/json-proxy", "workspace:packages/berry-json-proxy"],
             ["@berry/plugin-essentials", "workspace:packages/plugin-essentials"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["tmp", "npm:0.0.33"]
           ]
         }]
@@ -3843,7 +3843,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/json-proxy", "workspace:packages/berry-json-proxy"],
             ["@berry/parsers", "workspace:packages/berry-parsers"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["inquirer", "npm:6.2.1"],
             ["semver", "npm:5.6.0"],
             ["treeify", "npm:1.1.0"],
@@ -3904,7 +3904,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/core", "workspace:packages/berry-core"],
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/json-proxy", "workspace:packages/berry-json-proxy"],
-            ["clipanion", "npm:2.0.0-rc.10"]
+            ["clipanion", "npm:2.0.0-rc.15"]
           ]
         }]
       ]],
@@ -3940,7 +3940,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/plugin-npm", "workspace:packages/plugin-npm"],
             ["@berry/plugin-pack", "workspace:packages/plugin-pack"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["inquirer", "npm:6.2.1"],
             ["ssri", "npm:6.0.1"],
             ["yup", "npm:0.27.0"]
@@ -3956,7 +3956,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/core", "workspace:packages/berry-core"],
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@types/tar-stream", "npm:1.6.0"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["micromatch", "npm:4.0.2"],
             ["tar-stream", "npm:2.0.1"]
           ]
@@ -3972,7 +3972,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/plugin-stage", "workspace:packages/plugin-stage"],
             ["@berry/pnp", "workspace:packages/berry-pnp"],
-            ["clipanion", "npm:2.0.0-rc.10"]
+            ["clipanion", "npm:2.0.0-rc.15"]
           ]
         }]
       ]],
@@ -3984,7 +3984,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/builder", "workspace:packages/berry-builder"],
             ["@berry/core", "workspace:packages/berry-core"],
             ["@berry/fslib", "workspace:packages/berry-fslib"],
-            ["clipanion", "npm:2.0.0-rc.10"]
+            ["clipanion", "npm:2.0.0-rc.15"]
           ]
         }]
       ]],
@@ -4012,7 +4012,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/fslib", "workspace:packages/berry-fslib"],
             ["@berry/plugin-pack", "workspace:packages/plugin-pack"],
             ["@types/semver", "npm:5.5.0"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["semver", "npm:5.6.0"]
           ]
         }]
@@ -4026,7 +4026,7 @@ function $$SETUP_STATE(hydrateRuntimeState) {
             ["@berry/cli", "workspace:packages/berry-cli"],
             ["@berry/core", "workspace:packages/berry-core"],
             ["@berry/fslib", "workspace:packages/berry-fslib"],
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["p-limit", "npm:2.2.0"],
             ["yup", "npm:0.27.0"]
           ]
@@ -9166,10 +9166,10 @@ function $$SETUP_STATE(hydrateRuntimeState) {
         }]
       ]],
       ["clipanion", [
-        ["npm:2.0.0-rc.10", {
-          "packageLocation": "./.yarn/cache/clipanion-npm-2.0.0-rc.10-9894667b2939a9a81cad6ad38bbfa8361dba7d5273a645ea1dab1126c9a38f9c.zip/node_modules/clipanion/",
+        ["npm:2.0.0-rc.15", {
+          "packageLocation": "./.yarn/cache/clipanion-npm-2.0.0-rc.15-afda5a1ad1327c265d081e7cab734c7144dde5ddc42455564e742cbe6a83a3e1.zip/node_modules/clipanion/",
           "packageDependencies": [
-            ["clipanion", "npm:2.0.0-rc.10"],
+            ["clipanion", "npm:2.0.0-rc.15"],
             ["chalk", "npm:2.4.2"]
           ]
         }]
