+++ packages/berry-cli/package.json
@@ -12,7 +12,7 @@
     "@berry/parsers": "workspace:0.0.4",
     "@berry/shell": "workspace:0.0.3",
     "chalk": "^2.4.1",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "semver": "^5.6.0",
     "tmp": "^0.0.33",
     "yup": "^0.27.0"
