+++ packages/berry-core/package.json
@@ -13,7 +13,7 @@
     "agentkeepalive": "^4.0.2",
     "camelcase": "^5.3.1",
     "chalk": "^2.4.1",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "cross-spawn": "^6.0.5",
     "globby": "^8.0.1",
     "got": "^9.2.2",
