+++ packages/plugin-constraints/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "inquirer": "^6.2.0",
     "lodash.get": "^4.4.2",
     "node-emoji": "^1.8.1",
