+++ packages/plugin-dlx/package.json
@@ -8,7 +8,7 @@
     "@berry/fslib": "workspace:0.0.8",
     "@berry/json-proxy": "workspace:0.0.0",
     "@berry/plugin-essentials": "workspace:0.0.0",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "tmp": "^0.0.33"
   }
 }
