+++ packages/plugin-essentials/package.json
@@ -8,7 +8,7 @@
     "@berry/fslib": "workspace:0.0.8",
     "@berry/json-proxy": "workspace:0.0.0",
     "@berry/parsers": "workspace:0.0.4",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "inquirer": "^6.2.0",
     "semver": "^5.6.0",
     "treeify": "^1.1.0",
