+++ packages/plugin-init/package.json
@@ -6,6 +6,6 @@
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
     "@berry/json-proxy": "workspace:0.0.0",
-    "clipanion": "2.0.0-rc.10"
+    "clipanion": "2.0.0-rc.15"
   }
 }
