+++ packages/plugin-npm-cli/package.json
@@ -8,7 +8,7 @@
     "@berry/fslib": "workspace:0.0.8",
     "@berry/plugin-npm": "workspace:0.0.0",
     "@berry/plugin-pack": "workspace:0.0.0",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "inquirer": "^6.2.0",
     "ssri": "^6.0.1",
     "yup": "^0.27.0"
