+++ packages/plugin-pnp/package.json
@@ -7,7 +7,7 @@
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
     "@berry/pnp": "workspace:0.0.4",
-    "clipanion": "2.0.0-rc.10"
+    "clipanion": "2.0.0-rc.15"
   },
   "devDependencies": {
     "@berry/plugin-stage": "workspace:0.0.0"
