+++ packages/plugin-stage/package.json
@@ -6,7 +6,7 @@
     "@berry/builder": "workspace:0.0.0",
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
-    "clipanion": "2.0.0-rc.10"
+    "clipanion": "2.0.0-rc.15"
   },
   "scripts": {
     "build:plugin-stage": "run @berry-build-plugin"
