+++ packages/plugin-version/package.json
@@ -7,7 +7,7 @@
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
     "@berry/plugin-pack": "workspace:0.0.0",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "semver": "^5.6.0"
   },
   "devDependencies": {
