+++ packages/plugin-workspace-tools/package.json
@@ -7,7 +7,7 @@
     "@berry/cli": "workspace:2.0.0",
     "@berry/core": "workspace:0.0.6",
     "@berry/fslib": "workspace:0.0.8",
-    "clipanion": "2.0.0-rc.10",
+    "clipanion": "2.0.0-rc.15",
     "p-limit": "^2.2.0",
     "yup": "^0.27.0"
   },
